requirements.txt
python-dotenv
google-generativeai
pypdf
PyPDF2
chromadb
langchain
streamlit
langchain-google-genai 
langchain-community
-e .  
